requirements.txt
discord.py~=2.5.2
PyYAML~=6.0.2